requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
prometheus-client==0.21.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
Pillow==10.4.0
locust==2.29.0
fpdf2==2.8.7